feat(engine): add binding alias indirection for live re-exports; fix overflow and hash coverage in interpreter, interleaving explorer, and release checklist
module_live_binding.rs:
- Add BindingAlias struct for canonical live-binding indirection,
  enabling re-exported bindings to resolve to their source cell
  instead of a stale alias placeholder.
- Add LiveBindingMap::register_alias() for alias registration with
  deduplication and sorted insertion.
- Add resolve_binding_id() with cycle detection (BTreeSet visited)
  that walks the alias chain to the canonical source.
- Update get_cell(), get_cell_mut(), and update_value() to resolve
  through aliases before looking up cells.

baseline_interpreter.rs:
- Apply the same register overflow fix from the previous commit to
  two additional checked_add sites in the tail-call and multi-return
  paths, replacing misleading u32::MAX with the actual args.start
  register index.

interleaving_explorer.rs:
- Widen the coverage_millionths() intermediate product from i64 to
  i128 to prevent overflow when race_surfaces_covered is large,
  then truncate back to i64 for the result.

release_checklist_gate.rs:
- Include schema_version in the build_checklist_id hash computation,
  so checklist IDs change when the schema version changes — preventing
  cross-version collisions.

module_live_binding_integration.rs:
- Update re_export_binding_chain test to use register_alias() and
  wire imports through the alias rather than directly to the source.
- Add build_live_bindings_preserves_live_re_export_chain() integration
  test that constructs a full ModuleGraph → LiveBindingMap pipeline and
  verifies that updates to mod_a's binding propagate through mod_b's
  re-export to mod_c's import via alias resolution.

+++ crates/franken-engine/src/interleaving_explorer.rs
@@ -276,7 +276,7 @@ impl ExplorationReport {
         if self.race_surfaces_total == 0 {
             return 0;
         }
-        (self.race_surfaces_covered as i64 * 1_000_000) / self.race_surfaces_total as i64
+        (self.race_surfaces_covered as i128 * 1_000_000 / self.race_surfaces_total as i128) as i64
     }
 
     /// Number of failures found.
